+++ x/ecocredit/base/keeper/query_params_test.go
@@ -55,6 +55,9 @@ func TestQuery_Params(t *testing.T) {
 	})
 	assert.NilError(t, err)
 
+	allowedChain := "polygon"
+	assert.NilError(t, s.stateStore.AllowedBridgeChainTable().Insert(s.ctx, &api.AllowedBridgeChain{ChainName: allowedChain}))
+
 	result, err := s.k.Params(s.ctx, &types.QueryParamsRequest{})
 	assert.NilError(t, err)
 
@@ -64,4 +67,7 @@ func TestQuery_Params(t *testing.T) {
 	assert.Equal(t, result.Params.BasketFee.String(), sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(1000))).String())
 	assert.Equal(t, len(result.AllowedDenoms), 1)
 	assert.Equal(t, result.AllowedDenoms[0].BankDenom, "uregen")
+
+	assert.Equal(t, len(result.AllowedBridgeChains), 1)
+	assert.Equal(t, result.AllowedBridgeChains[0], allowedChain)
 }