+++ x/ecocredit/simulation/genesis.go
@@ -439,5 +439,11 @@ func genGenesisState(ctx context.Context, simState *module.SimulationState, ss a
 		return err
 	}
 
+	if err := ss.AllowedBridgeChainTable().Insert(ctx, &api.AllowedBridgeChain{
+		ChainName: "polygon",
+	}); err != nil {
+		return err
+	}
+
 	return nil
 }